Fix latin_hypercube dtype propagation; fix DummyAcquisition LSP violation
- Pass dtype=self.dtype to latin_hypercube() in _generate_points so
  LHS respects the runtime context dtype
- DummyAcquisition.score() adds missing cost_weighting parameter to
  match Acquisition base class signature (LSP compliance)

+++ src/activelearning/sampler/hypercube_sampler.py
@@ -108,10 +108,9 @@ class HypercubeSampler(Sampler):
         """
         n_dims = len(self.bounds)
         if self.point_strategy == "lhs":
-            return latin_hypercube(self.num_samples, n_dims)
+            return latin_hypercube(self.num_samples, n_dims, dtype=self.dtype)
         # Default: uniform
         return torch.rand(self.num_samples, n_dims, dtype=self.dtype)
-
 
     def _assign_fidelities(self) -> list[Optional[int]]:
         """Assign fidelity levels to ``num_samples`` candidates.
